move service to the standalone module
